requirements.txt
Django~=3.1
psycopg2-binary~=2.8
python-dotenv~=0.15
django-taggit~=1.3
Pillow~=8.1
django-mptt~=0.11
django-cleanup~=5.1
django-next-prev~=1.1